+++ src/main/scala/wdlTools/cli/Fix.scala
@@ -1,8 +1,7 @@
 package wdlTools.cli
 
-import java.nio.file.Paths
 import wdlTools.generators.code.Fixer
-import dx.util.FileSourceResolver
+import dx.util.{FileSourceResolver, FileUtils}
 
 import scala.language.reflectiveCalls
 
@@ -11,10 +10,12 @@ case class Fix(conf: WdlToolsConf) extends Command {
     val fileResolver = FileSourceResolver.get
     val fixer = Fixer(conf.fix.followImports(), fileResolver)
     val docSource = fileResolver.resolve(conf.fix.uri())
-    fixer.fix(docSource,
-              conf.fix.outputDir.toOption.getOrElse(Paths.get(".")),
-              conf.fix.overwrite(),
-              conf.fix.srcVersion.toOption,
-              Some(new CliTypeErrorHandler()))
+    fixer.fix(
+        docSource,
+        conf.fix.outputDir.toOption.getOrElse(FileUtils.cwd(absolute = true)),
+        conf.fix.overwrite(),
+        conf.fix.srcVersion.toOption,
+        Some(new CliTypeErrorHandler())
+    )
   }
 }