+++ RELEASE_NOTES.md
@@ -2,8 +2,9 @@
 
 ## in develop
 
-* Updates to dxCommon 0.4.2, which fixes the output of the `printTree` command
+* Updates dxCommon version, which fixes the output of the `printTree` command
 * Allow primitive -> `String` coercion for WDL versions <= 1.0
+* Fixes parsing of unary operations
 
 ## 0.14.1 (2021-06-08)